gcap devices were not displaying lines with line thicknesses between 1.0 and 2.0
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@90 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ncarview/src/lib/libctrans/gcapdev.c
@@ -1,5 +1,5 @@
 /*
- *	$Id: gcapdev.c,v 1.2 1991-03-12 18:46:02 clyne Exp $
+ *	$Id: gcapdev.c,v 1.3 1991-07-17 15:22:11 clyne Exp $
  */
 #include <stdio.h>
 #include "graphcap.h"
@@ -65,7 +65,7 @@ boolean polysim;	/* True if to simulate polygons with lines */
 	 * 	if line width is not the default (1) and device does not have
 	 *	wide line capability simulate wide lines with software
 	 */
-	if (!polyflag && LINE_WIDTH != 1 && LINE_WIDTH_START_SIZE == 0) {
+	if (!polyflag && (int) LINE_WIDTH != 1 && LINE_WIDTH_START_SIZE == 0) {
 		for(i=1;i<*coord_buf_num;i++)
 			(void) ComFatLine(coord_buf[i-1].x,coord_buf[i-1].y,
 					coord_buf[i].x,coord_buf[i].y,
